feat: FollowButtonをコンポーネントに
