+++ frontend/src/features/User/User.tsx
@@ -1,9 +1,7 @@
-import { chatSocketAtom } from '@/stores/auth';
 import { FTButton, FTH1, FTH4 } from '@/components/FTBasicComponents';
-import { useAction } from '@/hooks';
 import { useUpdateUser, useUserDataReadOnly } from '@/stores/store';
 import { useAtom } from 'jotai';
-import { Suspense, useEffect, useState } from 'react';
+import { Suspense, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import * as dayjs from 'dayjs';
 import { OnlineStatusDot } from '@/components/OnlineStatusDot';
@@ -14,76 +12,7 @@ import { APIError } from '@/errors/APIError';
 import { useManualErrorBoundary } from '@/components/ManualErrorBoundary';
 import { DmModal } from '../DM/DmModal';
 import { Modal } from '@/components/Modal';
-
-const FollowButton = (props: { userId: number; isFriend: boolean }) => {
-  const [mySocket] = useAtom(chatSocketAtom);
-  const command = {
-    follow: (targetId: number) => {
-      const data = {
-        userId: targetId,
-      };
-      console.log('ft_follow', data);
-      mySocket?.emit('ft_follow', data);
-    },
-    unfollow: (targetId: number) => {
-      const data = {
-        userId: targetId,
-      };
-      console.log('ft_unfollow', data);
-      mySocket?.emit('ft_unfollow', data);
-    },
-  };
-
-  type Phase = 'IsFriend' | 'IsNotFriend' | 'RunningUnfollow' | 'RunningFollow';
-  const [phase, setPhase] = useState<Phase>(
-    props.isFriend ? 'IsFriend' : 'IsNotFriend'
-  );
-  const [setIsRunning, isRunning] = useAction(false, (running) => {
-    console.log('setIsRunning', running);
-    if (!running) {
-      return;
-    }
-    if (props.isFriend) {
-      // Unfollow
-      setPhase('RunningUnfollow');
-      command.unfollow(props.userId);
-    } else {
-      // Follow
-      setPhase('RunningFollow');
-      command.follow(props.userId);
-    }
-  });
-
-  useEffect(() => {
-    if (props.isFriend) {
-      setPhase('IsFriend');
-    } else {
-      setPhase('IsNotFriend');
-    }
-    setIsRunning(false);
-  }, [props.isFriend]);
-
-  const text = (() => {
-    switch (phase) {
-      case 'IsFriend':
-        return 'Unfollow';
-      case 'IsNotFriend':
-        return 'Follow';
-      case 'RunningFollow':
-        return 'Following...';
-      case 'RunningUnfollow':
-        return 'Unfollowing...';
-    }
-  })();
-
-  return (
-    <>
-      <FTButton disabled={isRunning} onClick={() => setIsRunning(true)}>
-        {text}
-      </FTButton>
-    </>
-  );
-};
+import { FollowButton } from './components/FollowButton';
 
 const PresentatorView = (props: { personalData: TD.User }) => {
   const [friends] = useAtom(dataAtom.friends);
